Добавлены команды set и get

+++ cogs/__init__.py
@@ -7,11 +7,12 @@ import sys
 import asyncio
 import models
 import errors
+
 try:
-   import cPickle as pickle
+    import cPickle as pickle
 except ModuleNotFoundError:
-   import pickle
-from pprint import pprint
+    import pickle
+from collections import namedtuple
 
 langs = {}
 
@@ -32,6 +33,8 @@ def _(text, l='en'):
 
 
 init()
+Info = namedtuple('Info', ['server', 'server_settings', 'author', 'author_settings', 'user',
+                           'message'])
 
 
 class Commands:
@@ -40,27 +43,67 @@ class Commands:
 
     async def before_invoke(self, ctx):
         server = models.Guild.get_from_object(ctx.guild)
+        server[0].save()
         if ctx.author.bot:
             raise errors.IgnoringBot
-        author = models.User.get_from_object(ctx.author)
-        user = models.UserInGuild.get_from_object(author, server[0])
-        pprint(dir(ctx.message))
-        print(server, author, user)
+        user = models.User.get_from_object(ctx.author)
+        author = models.UserInGuild.get_from_object(user, server[0])
+        # message = models.Message.get_from_object(ctx.message, author[0])
+        message = None
+        return Info(server=server[0], server_settings=server[1], author=author[0],
+                    author_settings=author[1], user=user, message=message)
+
+    async def error(self):
+        blablabla
 
     async def ping(self, ctx):
         await self.before_invoke(ctx)
-        return await self.get_embed(_('Ping'), _('ping_command').format(latency=int(self.bot.latency*1000)))
+        return await self.get_embed(_('Ping'), _('ping_command').format(latency=int(self.bot.latency * 1000)))
+
+    async def get(self, ctx, setting: str):
+        info = await self.before_invoke(ctx)
+        setting = setting.lower()
+        s = info.server_settings['user'].get(setting)
+        if s is not None:
+            return await self.get_embed(_('Setting "{}"').format(setting), _('Value: **{}**').format(s))
+        else:
+            raise errors.SettingNotFound
+
+    async def set(self, ctx, setting: str, value: str):
+        info = await self.before_invoke(ctx)
+        setting = setting.lower()
+        schema = st.settings['schema'].get(setting)
+        if schema is not None:
+            info.server_settings['user'][setting] = schema['type'](value)
+            if schema['check_type'] == 'minmax':
+                if schema['max'] >= info.server_settings['user'][setting] >= schema['min']:
+                    info.server.save_settings(info.server_settings)
+                    return await self.get_embed(_('Success'), _('Applied!'))
+                else:
+                    info.server_settings['user'][setting] = schema['default']
+                    info.server.save_settings(info.server_settings)
+                    raise errors.MinMaxError
+            elif schema['check_type'] is None:
+                info.server.save_settings(info.server_settings)
+                return await self.get_embed(_('Success'), _('Applied!'))
+        else:
+            raise errors.SettingNotFound
 
     async def info(self, ctx):
-        await self.before_invoke(ctx)
+        info = await self.before_invoke(ctx)
         return await self.get_embed(_('Info'), _('command_info').format(commit=st.commit))
 
-    async def get_embed(self, title, description):
-        embed = discord.Embed(title=title, description=description)
+    async def get_embed(self, title, description, color='darkgrey'):
+        embed = discord.Embed(title=title, description=description, color=st.colors[color])
         embed.set_footer(text=_("bot_name"), icon_url=st.author_url)
         return embed
 
+    async def on_message(self, ctx):
+        await self.before_invoke(ctx)
+
     async def on_error(self, ctx, error):
+        info = await self.before_invoke(ctx)
+
         if hasattr(ctx.command, 'on_error'):
             return
 
@@ -86,9 +129,15 @@ class Commands:
             except discord.HTTPException:
                 pass
         elif isinstance(error, cog_errors.CommandNotFound):
+            if info.server_settings['user']['error_on_non_existent_command'] == 0:
+                return
             msg = _('Command not found')
+        elif isinstance(error, errors.SettingNotFound):
+            msg = _('Setting not found, read the docs.')
         elif isinstance(error, errors.IgnoringBot):
             return
+        elif isinstance(error, errors.MinMaxError):
+            msg = _('Min/Max error')
         elif isinstance(error, (discord.ext.commands.errors.MissingRequiredArgument, cog_errors.BadArgument)):
             msg = _('Check the arguments in the command')
         else:
@@ -96,8 +145,10 @@ class Commands:
             print('Ignoring exception in command {}:'.format(ctx.command), file=sys.stderr)
             traceback.print_exception(type(error), error, error.__traceback__, file=sys.stderr)
         m = await ctx.send(msg)
-        await asyncio.sleep(10)
-        await m.delete()
+        print(info.server_settings)
+        if info.server_settings['user']['delete_error_messages'] != 0:
+            await asyncio.sleep(info.server_settings['user']['delete_error_messages'])
+            await m.delete()
 
 
 def setup(bot):
